requirements.txt
slack-bolt
requests
python-dotenv